Move combo form elements
