+++ src/renderer/components/combos/CharacterSelect.tsx
@@ -4,7 +4,8 @@ import { Field } from "react-final-form";
 import Select, { components, MultiValueProps, OptionProps, OptionTypeBase, SingleValueProps } from "react-select";
 import styled from "styled-components";
 
-import { CharacterIcon } from "../../../components/CharacterIcon";
+import { CharacterIcon } from "../CharacterIcon";
+import { CharacterLabel } from "./CharacterLabel";
 
 export const sortedCharacterInfos: CharacterInfo[] = getAllCharacters()
   .sort((a, b) => {
@@ -24,24 +25,6 @@ const SingleValue: React.ComponentType<SingleValueProps<OptionTypeBase>> = (prop
 const MultiValueRemove: React.ComponentType<MultiValueProps<OptionTypeBase>> = (props) => {
   return (
     <components.MultiValueRemove {...props}><CharacterIcon character={props.data.value} /></components.MultiValueRemove>
-  );
-};
-
-const CharacterLabel: React.FC<{ characterId: Character, name: string, disabled?: boolean }> = (props) => {
-  const isDisabled = props.disabled;
-  const LabelContainer = styled.div`
-    display: flex;
-    align-items: center;
-    ${isDisabled && `
-      opacity: 0.5;
-      cursor: not-allowed;
-    `}
-  `;
-  return (
-    <LabelContainer>
-      <CharacterIcon character={props.characterId} grayscale={isDisabled} />
-      <span style={{marginLeft: "10px"}}>{props.name}</span>
-    </LabelContainer>
   );
 };
