+++ src/renderer/components/combos/PercentageSlider.tsx
